Move new hooks to the hooks directory.
